requirements.txt
shapely
faker
sqlalchemy
psycopg2-binary
python-dotenv
requests
geopy
shortuuid